requirements.txt
requests
pytest
openapi-python-client
python-dotenv
